Attempt at Dealing with Sensor Disconnect
This program is an attempt at dealing with a disconnect from the sensor to the Pi. The intention is to begin a new .csv file when an interruption is detected, and not record data when it's determined to still be disconnected.  However, it currently does not work. The sensor data read off the serial bus when the sensor is disconnected is the "null" character (a single character, rather than a string). The code continues to run without writing to .csv when the sensor is disconnected, but when the sensor is reconnected, it does not resume writing.